refactor legacy login screen and view model
