+++ feature/login/src/commonMain/kotlin/com/livefast/eattrash/raccoonforfriendica/feature/login/di/LoginModule.kt
@@ -1,13 +1,13 @@
 package com.livefast.eattrash.raccoonforfriendica.feature.login.di
 
-import com.livefast.eattrash.raccoonforfriendica.feature.login.LoginMviModel
-import com.livefast.eattrash.raccoonforfriendica.feature.login.LoginViewModel
+import com.livefast.eattrash.raccoonforfriendica.feature.login.legacy.LegacyLoginMviModel
+import com.livefast.eattrash.raccoonforfriendica.feature.login.legacy.LegacyLoginViewModel
 import org.koin.dsl.module
 
 val featureLoginModule =
     module {
-        factory<LoginMviModel> {
-            LoginViewModel(
+        factory<LegacyLoginMviModel> {
+            LegacyLoginViewModel(
                 credentialsRepository = get(),
                 apiConfigurationRepository = get(),
                 loginUseCase = get(),
