+++ feature/login/src/commonMain/kotlin/com/livefast/eattrash/raccoonforfriendica/feature/login/legacy/LegacyLoginMviModel.kt
@@ -1,4 +1,4 @@
-package com.livefast.eattrash.raccoonforfriendica.feature.login
+package com.livefast.eattrash.raccoonforfriendica.feature.login.legacy
 
 import androidx.compose.runtime.Stable
 import cafe.adriel.voyager.core.model.ScreenModel
@@ -6,9 +6,9 @@ import com.livefast.eattrash.raccoonforfriendica.core.architecture.MviModel
 import com.livefast.eattrash.raccoonforfriendica.core.utils.validation.ValidationError
 
 @Stable
-interface LoginMviModel :
+interface LegacyLoginMviModel :
     ScreenModel,
-    MviModel<LoginMviModel.Intent, LoginMviModel.State, LoginMviModel.Effect> {
+    MviModel<LegacyLoginMviModel.Intent, LegacyLoginMviModel.State, LegacyLoginMviModel.Effect> {
     sealed interface Intent {
         data class SetNodeName(
             val name: String,
